requirements.txt
hkp4py
pyyaml
